Make sure to always parse numbers as decimals in GeoJSON coordinates
By default the GraphQL library parses numbers from the incoming JSON
as Ints if they do not have a decimal dot in them. But the code
expects the coordinates to be decimals (graphql.language.FloatValue)
and performed a raw cast to that type. Which resulted in a
ClassCastException being thrown.

Code has been added to check whether the incoming value has been
parsed as IntValue or FloatValue and a proper conversion to Java
double is then performed.